Retrieve text to display using HTTP calls
Change the way the controller retrieves the values to be displayed.  An
HTTP call is now made to retrieve the text to be displayed, instead of
attempting to read from a file the web interface has written to.

The web interface currently provides the following endpoints for the
controller:

- /api/message gives the message of the day, as set by the user
- /api/currently_playing gives the current song and artist of the
  connected Spotify account
- /api/weather gives the current weather conditions

The controller currently displays the following sections during a full
iteration:

- The current time
- The message of the day
- The date
- The current weather conditions
- The current song and artist (if music is playing)